+++ projects/evo-ui-kit/src/lib/components/evo-navigation-tabs/public-api.ts
@@ -1,4 +1,5 @@
 export * from './evo-navigation-tabs.module';
 export * from './evo-navigation-tabs.component';
 export * from './evo-navigation-tab.directive';
-export * from "./types/evo-tabs-size";
+export * from './types/evo-navigation-tabs-size';
+export * from './interfaces/evo-navigation-tab';