+++ projects/evo-ui-kit/src/lib/components/evo-tooltip/evo-tooltip.module.ts
@@ -1,0 +1,12 @@
+import {NgModule} from '@angular/core';
+import {CommonModule} from '@angular/common';
+import {EvoTooltipComponent} from './evo-tooltip.component';
+import {EvoTooltipDirective} from './directives/evo-tooltip.directive';
+import {OverlayModule} from '@angular/cdk/overlay';
+
+@NgModule({
+    imports: [OverlayModule, CommonModule],
+    declarations: [EvoTooltipComponent, EvoTooltipDirective],
+    exports: [EvoTooltipDirective],
+})
+export class EvoTooltipModule {}