+++ projects/evo-ui-kit/src/lib/components/evo-tooltip/interfaces/evo-tooltip-styles.ts
@@ -1,0 +1,5 @@
+import {EvoTooltipVariableArrowPosition} from '../enums/evo-tooltip-variable-arrow-position';
+
+export type EvoTooltipStyles = {
+    [key in EvoTooltipVariableArrowPosition]: string;
+};